mudanca no formato moeda de cadastro do valor, e criacao do alguns mu…

+++ projeto-individual/database/migrations/2022_06_29_110929_create_releases_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('release_type');
             $table->string('person');
             $table->string('description');
-            $table->double('amount', 16,2);
+            $table->decimal('amount', 12,2);
             $table->date('due_date');
             $table->timestamps();
         });
